requirements.txt
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
constantly==15.1.0
cryptography==40.0.1
daphne==3.0.2
Django==4.2
django-cors-headers==3.14.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
h2==4.1.0
hiredis==2.2.2
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
lxml==4.9.2
msgpack==1.0.5
numpy==1.24.2
priority==1.3.0
psutil==5.9.4
psycopg-binary==3.1.8
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
zope.interface==6.0
